fix: The delete sector button must be in the content of the sector tab
